fixing php errors in demos, games, interviews

+++ Website/AtariLegend/admin/demos/demos_upload.php
@@ -17,13 +17,11 @@
 
 //load all common functions
 include("../includes/common.php"); 
-include("../includes/config.php"); 
-
 
 //****************************************************************************************
 // We wanna add a new download
 //**************************************************************************************** 
-if ($action == 'add_download' )
+if (isset($action) and $action == 'add_download' )
 {
 	require_once('../includes/pclzip.lib.php');
 	
@@ -116,7 +114,7 @@ if ($action == 'add_download' )
 //****************************************************************************************
 // When the update button has been pressed, the file name and comments get updated
 //**************************************************************************************** 
-if ($action == 'update_download')
+if (isset($action) and $action == 'update_download')
 {
 	mysql_query("UPDATE demo_download SET cracker='$cracker', supplier='$supplier', screen='$screen', language='$language', trainer='$trainer', legend='$legend', disks='$disks', set_nr='$set_nr', intro='$intro', harddrive='$harddrive', disable='$disable', version='$version', tos='$tos' WHERE demo_download_id='$demo_download_id'");
 }
@@ -124,7 +122,7 @@ if ($action == 'update_download')
 //****************************************************************************************
 // This is where the file name and comments get deleted
 //**************************************************************************************** 
-if ($action == "delete_download")
+if (isset($action) and $action == "delete_download")
 {
 	mysql_query("DELETE from demo_download WHERE demo_download_id='$demo_download_id'");
 	unlink ("$demo_file_path$demo_download_id.zip");
@@ -146,54 +144,56 @@ $demo_info=mysql_fetch_array ($SQL_DEMO_INFO);
 //get some basic demo info
 $smarty->assign('demo',
 		 array('demo_id' => $demo_id,
-		 	   'demo_name' => $demo_info[demo_name]));
+		 	   'demo_name' => $demo_info['demo_name']));
 
 //get the existing downloads
 $SQL_DOWNLOADS = mysql_query("SELECT * FROM demo_download WHERE demo_id='$demo_id'")
 			  	 or die ("Error getting download info");		
+
+$nr_downloads = 1;
 
 while ($downloads=mysql_fetch_array($SQL_DOWNLOADS)) 
 {
 	// first lets create the filenames
 	$filename = "$demo_info[demo_name]";
 	
-	if ($demo_info[demo_year]=='') { $filename .= " (19xx)"; } else { $filename .= " ($demo_info[demo_year])";}
-	if ($demo_info[crew_id]!=='') { $filename .= "($demo_info[crew_name])"; }
-	if ($downloads[demo_ext]=="stx") { $filename .="[pasti]";}
-	if ($downloads[demo_ext]=="msa") { $filename .="[MSA]";}
-	if ($downloads[cracker]!=="") { $filename .="[cr $downloads[cracker]]";}
-	if ($downloads[supplier]!=="") { $filename .="[su $downloads[supplier]]";}
-	if ($downloads[screen]!=="") { $filename .="[$downloads[screen]]";}
-	if ($downloads[language]!=="") { $filename .="[$downloads[language]]";}
-	if ($downloads[trainer]!=="") { $filename .="[$downloads[trainer]]";}
-	if ($downloads[legend]=="1") { $filename .="[AL]";}
-	if ($downloads[disks]!=="0" || $downloads[disks]!=="") { $filename .=" Disk $downloads[disks]";}
+	if ($demo_info['demo_year']=='') { $filename .= " (19xx)"; } else { $filename .= " ($demo_info[demo_year])";}
+	if ($demo_info['crew_id']!=='') { $filename .= "($demo_info[crew_name])"; }
+	if ($downloads['demo_ext']=="stx") { $filename .="[pasti]";}
+	if ($downloads['demo_ext']=="msa") { $filename .="[MSA]";}
+	if ($downloads['cracker']!=="") { $filename .="[cr $downloads[cracker]]";}
+	if ($downloads['supplier']!=="") { $filename .="[su $downloads[supplier]]";}
+	if ($downloads['screen']!=="") { $filename .="[$downloads[screen]]";}
+	if ($downloads['language']!=="") { $filename .="[$downloads[language]]";}
+	if ($downloads['trainer']!=="") { $filename .="[$downloads[trainer]]";}
+	if ($downloads['legend']=="1") { $filename .="[AL]";}
+	if ($downloads['disks']!=="0" || $downloads['disks']!=="") { $filename .=" Disk $downloads[disks]";}
 	$filename .=".zip";
 	
 	$filepath = $demo_file_path;
-	$filepath .= $downloads[demo_download_id];
+	$filepath .= $downloads['demo_download_id'];
 	
 	//convert the date
 	$date = convert_timestamp($downloads['date']);
 	
 	//start filling the smarty object
 	$smarty->append('downloads',
-	    	 array('demo_download_id' => $downloads[demo_download_id],
-			 	   'cracker' => $downloads[cracker],
-				   'supplier' => $downloads[supplier],
-				   'intro' => $downloads[intro],
-				   'harddrive' => $downloads[harddrive],
-				   'pal_ntsc' => $downloads[screen],
-				   'language' => $downloads[language],
-				   'trainer' => $downloads[trainer],
-				   'disks' => $downloads[disks],
-				   'set_nr' => $downloads[set_nr],
-				   'legend' => $downloads[legend],
-				   'disable' => $downloads[disable],
+	    	 array('demo_download_id' => $downloads['demo_download_id'],
+			 	   'cracker' => $downloads['cracker'],
+				   'supplier' => $downloads['supplier'],
+				   'intro' => $downloads['intro'],
+				   'harddrive' => $downloads['harddrive'],
+				   'pal_ntsc' => $downloads['screen'],
+				   'language' => $downloads['language'],
+				   'trainer' => $downloads['trainer'],
+				   'disks' => $downloads['disks'],
+				   'set_nr' => $downloads['set_nr'],
+				   'legend' => $downloads['legend'],
+				   'disable' => $downloads['disable'],
 				   'filename' => $filename,
 				   'filepath' => $filepath,
-				   'version' => $downloads[version],
-				   'tos' => $downloads[tos],
+				   'version' => $downloads['version'],
+				   'tos' => $downloads['tos'],
 				   'date' => $date));
 
 	$nr_downloads++;
@@ -201,7 +201,7 @@ while ($downloads=mysql_fetch_array($SQL_DOWNLOADS))
 
 $smarty->assign('nr_downloads',$nr_downloads);
 
-$smarty->assign("user_id",$_SESSION[user_id]);
+$smarty->assign("user_id",$_SESSION['user_id']);
 $smarty->assign('demos_upload_tpl', '1');
 
 //Send all smarty variables to the templates
